+++ src/gap.h
@@ -189,11 +189,22 @@ void gap_set_bondable_mode(int enabled);
 int gap_get_bondable_mode(void);
 
 /**
+ * @brief Set security level for all outgoing and incoming connections. Default: LEVEL_2
+ * @param security_level
+ * @note has to be called before services or profiles are initialized
+ */
+void gap_set_security_level(gap_security_level_t security_level);
+
+/**
+ * @brief Get security level
+ * @return security_level
+ */
+gap_security_level_t gap_get_security_level(void);
+
+/**
  * @brief Register filter for rejecting classic connections. Callback will return 1 accept connection, 0 on reject.
  */
-
 void gap_register_classic_connection_filter(int (*accept_callback)(bd_addr_t addr));
-
 
 /* Configure Secure Simple Pairing */